V4/speed widget update (#597)
* Speed widgets
