+++ modules/speed/SpeedGraph.tsx
@@ -21,6 +21,8 @@ import type { SpeedDataPoint } from '../../common/types/dataPoints';
 import { drawSimpleTitle } from '../../common/components/charts/Title';
 import { useBreakpoint } from '../../common/hooks/useBreakpoint';
 import { watermarkLayout } from '../../common/constants/charts';
+import { ChartBorder } from '../../common/components/charts/ChartBorder';
+import { ChartDiv } from '../../common/components/charts/ChartDiv';
 import { CORE_TRACK_LENGTHS, PEAK_MPH } from './constants/speeds';
 import type { ParamsType } from './constants/speeds';
 
@@ -57,123 +59,127 @@ export const SpeedGraph: React.FC<SpeedGraphProps> = ({
   const isMobile = !useBreakpoint('md');
   const labels = data.map((point) => point.date);
   return (
-    <Line
-      id={'Speed'}
-      height={240}
-      ref={ref}
-      redraw={true}
-      data={{
-        labels,
-        datasets: [
-          {
-            label: `MPH`,
-            borderColor: LINE_COLORS[line ?? 'default'],
-            pointRadius: 8,
-            pointBackgroundColor: 'transparent',
-            pointBorderWidth: 0,
-            stepped: true,
-            pointHoverRadius: 3,
-            spanGaps: false,
-            pointHoverBackgroundColor: LINE_COLORS[line ?? 'default'],
-            data: data.map((datapoint) =>
-              datapoint.value
-                ? (CORE_TRACK_LENGTHS[line ?? 'DEFAULT'] / (datapoint.value / 3600)).toFixed(1)
-                : Number.NaN
-            ),
-          },
-        ],
-      }}
-      options={{
-        responsive: true,
-        maintainAspectRatio: false,
-        layout: {
-          padding: {
-            top: showTitle ? 25 : 0,
-          },
-        },
-        interaction: {
-          intersect: false,
-        },
-        // @ts-expect-error The watermark plugin doesn't have typescript support
-        watermark: watermarkLayout(isMobile),
-        plugins: {
-          tooltip: {
-            mode: 'index',
-            position: 'nearest',
-            callbacks: callbacks,
-          },
-          legend: {
-            display: false,
-          },
-          title: {
-            // empty title to set font and leave room for drawTitle fn
-            display: showTitle,
-            text: '',
-          },
-        },
-        scales: {
-          y: {
-            suggestedMin: 0,
-            suggestedMax: PEAK_MPH[line ?? 'DEFAULT'],
-            display: true,
-            ticks: {
-              color: COLORS.design.subtitleGrey,
-            },
-            title: {
-              display: true,
-              text: 'mph',
-              color: COLORS.design.subtitleGrey,
-            },
-          },
-          x: {
-            min: startDate,
-            max: endDate,
-            type: 'time',
-            time: {
-              unit: unit,
-              tooltipFormat: tooltipFormat,
-              displayFormats: {
-                month: 'MMM',
+    <ChartBorder>
+      <ChartDiv isMobile={isMobile}>
+        <Line
+          id={'Speed'}
+          height={isMobile ? 240 : 200}
+          ref={ref}
+          redraw={true}
+          data={{
+            labels,
+            datasets: [
+              {
+                label: `MPH`,
+                borderColor: LINE_COLORS[line ?? 'default'],
+                pointRadius: 8,
+                pointBackgroundColor: 'transparent',
+                pointBorderWidth: 0,
+                stepped: true,
+                pointHoverRadius: 3,
+                spanGaps: false,
+                pointHoverBackgroundColor: LINE_COLORS[line ?? 'default'],
+                data: data.map((datapoint) =>
+                  datapoint.value
+                    ? (CORE_TRACK_LENGTHS[line ?? 'DEFAULT'] / (datapoint.value / 3600)).toFixed(1)
+                    : Number.NaN
+                ),
+              },
+            ],
+          }}
+          options={{
+            responsive: true,
+            maintainAspectRatio: false,
+            layout: {
+              padding: {
+                top: showTitle ? 25 : 0,
               },
             },
-            ticks: {
-              color: COLORS.design.subtitleGrey,
+            interaction: {
+              intersect: false,
             },
-            adapters: {
-              date: {
-                locale: enUS,
+            // @ts-expect-error The watermark plugin doesn't have typescript support
+            watermark: watermarkLayout(isMobile),
+            plugins: {
+              tooltip: {
+                mode: 'index',
+                position: 'nearest',
+                callbacks: callbacks,
+              },
+              legend: {
+                display: false,
+              },
+              title: {
+                // empty title to set font and leave room for drawTitle fn
+                display: showTitle,
+                text: '',
               },
             },
-            display: true,
-            title: {
-              display: false,
-              text: ``,
+            scales: {
+              y: {
+                suggestedMin: 0,
+                suggestedMax: PEAK_MPH[line ?? 'DEFAULT'],
+                display: true,
+                ticks: {
+                  color: COLORS.design.subtitleGrey,
+                },
+                title: {
+                  display: true,
+                  text: 'mph',
+                  color: COLORS.design.subtitleGrey,
+                },
+              },
+              x: {
+                min: startDate,
+                max: endDate,
+                type: 'time',
+                time: {
+                  unit: unit,
+                  tooltipFormat: tooltipFormat,
+                  displayFormats: {
+                    month: 'MMM',
+                  },
+                },
+                ticks: {
+                  color: COLORS.design.subtitleGrey,
+                },
+                adapters: {
+                  date: {
+                    locale: enUS,
+                  },
+                },
+                display: true,
+                title: {
+                  display: false,
+                  text: ``,
+                },
+              },
             },
-          },
-        },
-      }}
-      plugins={[
-        {
-          id: 'customTitle',
-          afterDraw: (chart) => {
-            if (!data) {
-              // No data is present
-              const { ctx } = chart;
-              const { width } = chart;
-              const { height } = chart;
-              chart.clear();
+          }}
+          plugins={[
+            {
+              id: 'customTitle',
+              afterDraw: (chart) => {
+                if (!data) {
+                  // No data is present
+                  const { ctx } = chart;
+                  const { width } = chart;
+                  const { height } = chart;
+                  chart.clear();
 
-              ctx.save();
-              ctx.textAlign = 'center';
-              ctx.textBaseline = 'middle';
-              ctx.font = "16px normal 'Helvetica Nueue'";
-              ctx.fillText('No data to display', width / 2, height / 2);
-              ctx.restore();
-            }
-            if (showTitle) drawSimpleTitle(`Median Speed`, chart);
-          },
-        },
-      ]}
-    />
+                  ctx.save();
+                  ctx.textAlign = 'center';
+                  ctx.textBaseline = 'middle';
+                  ctx.font = "16px normal 'Helvetica Nueue'";
+                  ctx.fillText('No data to display', width / 2, height / 2);
+                  ctx.restore();
+                }
+                if (showTitle) drawSimpleTitle(`Median Speed`, chart);
+              },
+            },
+          ]}
+        />
+      </ChartDiv>
+    </ChartBorder>
   );
 };
