+++ modules/speed/SpeedWidget.tsx
@@ -6,8 +6,8 @@ import { ChartPlaceHolder } from '../../common/components/graphics/ChartPlaceHol
 import { WidgetDiv } from '../../common/components/widgets/WidgetDiv';
 import { OVERVIEW_OPTIONS, TODAY_STRING } from '../../common/constants/dates';
 import { HomescreenWidgetTitle } from '../dashboard/HomescreenWidgetTitle';
-import { SpeedGraphWrapper } from './SpeedOverviewWrapper';
 import { getSpeedGraphConfig } from './constants/speeds';
+import { SpeedGraphWrapper } from './SpeedGraphWrapper';
 
 export const SpeedWidget: React.FC = () => {
   const { line, query } = useDelimitatedRoute();
